엔티티 수정 및 개인 파일 추가

+++ fulfillment/src/main/java/com/springboot/fulfillment/data/dto/OrdersDTO.java
@@ -33,7 +33,7 @@ public class OrdersDTO {
 	private String customerDetailAddress;
 	// 판매 상품
 	private Long goodsNo;
-	private Integer goodsCode;
+	private Long goodsCode;
 	private String goodsName;
 	private String goodsDescription;
 	private String goodsImg1;
